+++ MyEngine/SourceEditor/Editor_ImGui/EditorMainFrame_ImGui.cpp
@@ -914,13 +914,13 @@ void EditorMainFrame_ImGui::DrawGameAndEditorWindows(EngineCore* pEngineCore)
                 light1.m_Attenuation.Set( 5, 0.1f, 0.01f ); // x is a radius/range, yz not used.
                 light1.m_Color.Set( 1, 1, 1, 1 );
                 light1.m_LightType = LightType_Point;
-                light1.m_Position.Set( 2*cos(time), 1, 2*sin(time) );
+                light1.m_Position.Set( 2*cosf(time), 1, 2*sinf(time) );
 
                 MyLight light2;
                 light2.m_Attenuation.Set( 5, 0.1f, 0.01f ); // x is a radius/range, yz not used.
                 light2.m_Color.Set( 1, 1, 1, 1 );
                 light2.m_LightType = LightType_Point;
-                light2.m_Position.Set( 2*cos(PI+time), 1, 2*sin(PI+time) );
+                light2.m_Position.Set( 2*cosf(PI+time), 1, 2*sinf(PI+time) );
 
                 MyLight* lights[] = { &light1, &light2 };
                 pMeshBall->Draw( &matproj, &matview, nullptr, &campos, &camrot, lights, 2, nullptr, nullptr, nullptr, nullptr );
